Refactor helper directory to following a "separation of concerns" strategy.
