+++ backend/src/businessService/todosService.ts
@@ -1,5 +1,5 @@
-import { ToDosAccess } from "./todosAcess";
-import { AttachmentUtils } from "./attachmentUtils";
+import { ToDosAccess } from "../dataAccessLayer/todosAccess";
+import { AttachmentUtils } from "../utils/attachmentUtils";
 import { TodoItem } from "../models/TodoItem";
 import { CreateTodoRequest } from "../requests/CreateTodoRequest";
 import { UpdateTodoRequest } from "../requests/UpdateTodoRequest";
