+++ backend/src/lambda/http/createTodo.ts
@@ -4,7 +4,7 @@ import * as middy from 'middy'
 import { cors } from 'middy/middlewares'
 import { CreateTodoRequest } from '../../requests/CreateTodoRequest'
 import { getUserId } from '../utils';
-import { createToDo } from '../../helpers/todos'
+import { createToDo } from '../../businessService/todosService'
 import { createLogger } from '../../utils/logger'
 
 const logger = createLogger('CreateTODO')
