+++ backend/src/utils/attachmentUtils.ts
